​​refactor: 타입들 분리
