+++ src/types/common.ts
@@ -1,0 +1,16 @@
+export type SortOption = '높은 가격순' | '낮은 가격순';
+export type CategoryOption = '전체' | '패션잡화' | '식료품';
+
+export type Product = {
+  id: number;
+  name: string;
+  price: number;
+  imageUrl: string;
+  category: string;
+};
+
+export type CartItem = {
+  id: number;
+  product: Product;
+  quantity: number;
+};
